requirements.txt
pyodbc
alembic
apscheduler
fastapi
uvicorn
python-dotenv
pytelegrambotapi
sqlalchemy
oauth2client
google-api-python-client
